backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
gunicorn
whitenoise
